refactor(app): refactor archive page
