+++ src/modules/comment/index.tsx
@@ -1,32 +1,32 @@
-"use client";
+"use client"
 
-import { useOptimistic } from "react";
-import { CommentList } from "./commentList";
-import CommentForm from "./commentForm";
-import { postComment } from "@/app/(grid)/archives/[id]/serverActions";
-import { CommentPayload, CommentsPayload } from "@/db/query";
-import { Container } from "@yamada-ui/react";
+import { useOptimistic } from "react"
+import { CommentList } from "./commentList"
+import CommentForm from "./commentForm"
+import { CommentPayload, CommentsPayload } from "@/db/query"
+import { Container } from "@yamada-ui/react"
+import { postComment } from "@/app/archives/[id]/serverActions"
 
 export default function Comment(props: {
-    comments: CommentsPayload;
-    copipe_id: number;
+  comments: CommentsPayload
+  copipe_id: number
 }) {
-    const { comments, copipe_id } = props;
-    const [optimisticComments, addOptimisticComment] = useOptimistic(
-        comments,
-        (state, newComent: CommentPayload) => {
-            return [...state, newComent];
-        },
-    );
+  const { comments, copipe_id } = props
+  const [optimisticComments, addOptimisticComment] = useOptimistic(
+    comments,
+    (state, newComent: CommentPayload) => {
+      return [...state, newComent]
+    },
+  )
 
-    return (
-        <Container>
-            <CommentList comments={optimisticComments} />
-            <CommentForm
-                copipe_id={copipe_id}
-                addOptimisticComment={addOptimisticComment}
-                postComment={postComment}
-            />
-        </Container>
-    );
+  return (
+    <Container>
+      <CommentList comments={optimisticComments} />
+      <CommentForm
+        copipe_id={copipe_id}
+        addOptimisticComment={addOptimisticComment}
+        postComment={postComment}
+      />
+    </Container>
+  )
 }
